requirements.txt
aioconsole==0.4.1
ansible==6.0.0
ansible-core==2.13.1
anyio==3.5.0
appnope==0.1.3
asgiref==3.5.0
asn1crypto==1.5.1
asttokens==2.0.5
asyncio==3.4.3
attrs==21.4.0
backcall==0.2.0
backports.cached-property==1.0.1
black==22.3.0
bytewax==0.10.0
cbor2==5.4.3
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
coloredlogs==15.0.1
confluent-kafka==1.8.2
cryptography==36.0.2
curio==1.5
datadog==0.44.0
decorator==5.1.1
dill==0.3.5.1
docker==5.0.3
executing==0.8.3
Faker==13.15.1
fastapi==0.75.1
flake8==4.0.1
future==0.18.2
GeoAlchemy2==0.12.1
gevent==21.12.0
graphql-core==3.2.1
greenlet==1.1.2
grpcio==1.46.3
grpcio-tools==1.46.3
h11==0.13.0
h2==4.1.0
hpack==4.0.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.3
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
msgpack==1.0.3
msgpack-python==0.5.6
multidict==6.0.2
multiprocess==0.70.13
mypy==0.960
mypy-extensions==0.4.3
neo4j==4.4.4
noiseprotocol==0.3.1
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pydantic-factories==1.4.1
pyflakes==2.4.0
Pygments==2.11.2
pymgclient==1.2.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-mock==3.7.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.28.0
resolvelib==0.8.1
Shapely==1.8.2
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.34
sqlalchemy2-stubs==0.0.2a21
sqlmodel==0.0.6
stack-data==0.2.0
starlette==0.17.1
strawberry-graphql==0.105.0
toml==0.10.2
tomli==2.0.1
traitlets==5.2.1.post0
typer==0.4.1
types-toml==0.10.7
typing_extensions==4.1.1
ulid==1.1
ulid-py==1.1.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
wcwidth==0.2.5
webauthn==1.5.2
websocket==0.2.1
websocket-client==1.3.3
websockets==10.3
xeger==0.3.5
zerorpc==0.6.3
zope.event==4.5.0
zope.interface==5.4.0
